:white_check_mark: Move table tests to markdown_extensions
